Revert "Extract IosPushNotificationAppDelegate into a separate class"
This reverts commit 38a774eba96bd3b8127a4a5b802d88f5cbed61fa.

# Conflicts:
#	Softeq.XToolkit.PushNotifications.Droid/Softeq.XToolkit.PushNotifications.Droid.csproj
